Moved around the locations of a lot of our `parse` functions
* Began work on synthesis uniques/bases. Conversion works, but
  some synthesis-item specific mods are currently not sanitized.
